refactor(handlers): move namedmaphandler and namedpairhandler to new files
